Refactor of the package names
